code reorganization: moved doc strings to separate header
improves readability of main API definition
